start unifying group and main DB
