+++ src/lib/database/database.py
@@ -1,13 +1,18 @@
 from dataclasses import dataclass, field
 from typing import Protocol
+from uuid import UUID
+
 import pandas as pd
 from sqlalchemy.future import Engine
+from sqlmodel import Session, SQLModel, col, create_engine, select
+
 from src.lib.constants import DATABASE_PATH_TMP
-from sqlmodel import Session, SQLModel, create_engine, select, col
+from src.lib.groups import Group, GroupType
 from src.lib.models.data import *
+from src.lib.models.group import GroupDBModel
 
 
-class MainDB(Protocol):
+class Database(Protocol):
     def get_metadata(self, ms_ids: list[str]) -> pd.DataFrame:
         """One stop shop for simple search/SELECT queries.
 
@@ -61,6 +66,30 @@ class MainDB(Protocol):
 
     def txt_lookup_list(self) -> list[str]:
         """Returns the lookup-list for the texts. Used in front end search."""
+        ...
+
+    def get_ms_groups(self) -> list[Group]:
+        """Gets all groups of type `ManuscriptGroup` from the database."""
+        ...
+
+    def get_ppl_groups(self) -> list[Group]:
+        """Gets all groups of type `PeopleGroup` from the database."""
+        ...
+
+    def get_txt_groups(self) -> list[Group]:
+        """Gets all groups of type `TextGroup` from the database."""
+        ...
+
+    def get_all_groups(self) -> list[Group]:
+        """Gets all groups from the database."""
+        ...
+
+    def add_group(self, group: Group) -> None:
+        """Adds a new group to the database"""
+        ...
+
+    def update_group(self, group: Group, group_id: UUID) -> None:
+        """Updates a group in the database, either replacing its previous version, or creating it anew."""
         ...
 
 
@@ -141,3 +170,42 @@ class MainDBImpl:
         with Session(self.engine) as session:
             res = session.exec(select(TextDBModel.text_id)).all()
             return res
+
+    def get_ms_groups(self) -> list[Group]:
+        with Session(self.engine) as session:
+            statement = select(GroupDBModel).where(GroupDBModel.group_type == GroupType.ManuscriptGroup)
+            groups = session.exec(statement)
+            return [g.to_group() for g in groups]
+
+    def get_ppl_groups(self) -> list[Group]:
+        with Session(self.engine) as session:
+            statement = select(GroupDBModel).where(GroupDBModel.group_type == GroupType.PersonGroup)
+            groups = session.exec(statement)
+            return [g.to_group() for g in groups]
+
+    def get_txt_groups(self) -> list[Group]:
+        with Session(self.engine) as session:
+            statement = select(GroupDBModel).where(GroupDBModel.group_type == GroupType.TextGroup)
+            groups = session.exec(statement)
+            return [g.to_group() for g in groups]
+
+    def get_all_groups(self) -> list[Group]:
+        with Session(self.engine) as session:
+            statement = select(GroupDBModel)
+            groups = session.exec(statement)
+            return [g.to_group() for g in groups]
+
+    def add_group(self, group: Group) -> None:
+        with Session(self.engine) as session:
+            db_model = GroupDBModel.make(group)
+            session.add(db_model)
+            session.commit()
+
+    def update_group(self, group: Group, group_id: UUID) -> None:
+        with Session(self.engine) as session:
+            group_old = session.get(GroupDBModel, group_id)
+            if group_old is not None:
+                session.delete(group_old)
+            group_new = GroupDBModel.make(group)
+            session.add(group_new)
+            session.commit()
